Fix warnings about too big matrix size

+++ matrixmult/helloDSP/dsp/helloDSP_config.h
@@ -1,7 +1,7 @@
 /** ============================================================================
  *  @file   helloDSP_config.h
  *
- *  @path   
+ *  @path
  *
  *  @desc   Header file for MSGQ and POOL configurations for helloDSP.
  *
@@ -50,12 +50,12 @@ extern "C" {
 
 /* Control message data structure. */
 /* Must contain a reserved space for the header */
-typedef struct ControlMsg 
+typedef struct ControlMsg
 {
     MSGQ_MsgHeader header;
     Uint16 command;
     Char arg1[ARG_SIZE];
-	int mat[MAX_MATSIZE][MAX_MATSIZE];
+	  Uint16 mat[MAX_MATSIZE][MAX_MATSIZE];
 } ControlMsg;
